requirements-client.txt
requests
python-dotenv
libcst
openai
fasttext
scipy
levenshtein
pytest
pytest-reportlog
autoimport
transformers
coverage